mediaQuery

+++ src/components/pages/portfolio/portfolio.tsx
@@ -1,14 +1,11 @@
+import React, { useState } from 'react';
 import { observer } from 'mobx-react-lite';
-import React, { useState, useCallback } from 'react';
 import authStore from '../../../store/authStore';
 import portfolioStore from '../../../store/portfolioStore';
-import ErrorMessage from '../../popupMessages/showMessage';
 import ShowMessage from '../../popupMessages/showMessage';
-// import PriceGrid from './priceGrid';
-// import PriceForm from './priceForm';
-// import EditMenu from '../../ui/editMenu';
 import PortfolioGrid from './portfolioGrid';
-import PortfolioPanel from './PortfolioPanel';
+import PortfolioPanel from '../../ui/EditPanel';
+import useConfirm from '../../../hooks/useConfirm';
 
 
 export interface Event<T = EventTarget> {
@@ -25,13 +22,10 @@ interface IPortfolioProps {
 const Portfolio: React.FC<IPortfolioProps> = observer(() => {
   const { isAuth } = authStore;
   const { uploadImages, deleteAllImages, portfolioError, portfolioSuccess } = portfolioStore;
-  // const [formOpen, setFormOpen] = React.useState<boolean>(false);
-  // const [menuActionType, setMenuActionType] = React.useState<MenuActionType>('ADD');
-
-  const [images, setImages] = useState<any>(null);
+  const [ images, setImages ] = useState<any>(null);
+  const { confirm } = useConfirm();
 
   const changeHandler = (e: Event<HTMLInputElement>) => {
-    console.log('changeHandler')
     const data = new FormData();
     const files = { ...e.target.files };
     e.target.files = null;
@@ -43,18 +37,22 @@ const Portfolio: React.FC<IPortfolioProps> = observer(() => {
     setImages(data);
   }
 
-  const uploadHandler = () => {
-    console.log('uploadHandler')
+  const addHandler = () => {
     uploadImages(images);
     setImages(null);
   }
 
-  const removeAllHandler = () => {
-    console.log('removeAllHandler')
-    deleteAllImages();
-    setImages(null);
+  const showConfirm = async () => {
+    const isConfirmed = await confirm('Удалить все записи?');
+    if(isConfirmed) {
+      deleteAllImages();
+      setImages(null);
+    }
   }
 
+  const removeAllHandler = () => {
+    showConfirm();  
+  }
 
   return (
     <>
@@ -62,32 +60,16 @@ const Portfolio: React.FC<IPortfolioProps> = observer(() => {
         error={portfolioError}
         success={portfolioSuccess}
       />
+
       <PortfolioGrid />
 
       {isAuth &&
         <PortfolioPanel
           changeHandler={changeHandler}
-          uploadHandler={uploadHandler}
+          addHandler={addHandler}
           removeAllHandler={removeAllHandler}
         />
       }
-
-      {/* <ImageGrid
-        setFormOpen={setFormOpen}
-        menuActionType={menuActionType}
-        setMenuActionType={setMenuActionType}
-      /> */}
-      {/* <PriceForm
-        formOpen={formOpen}
-        setFormOpen={setFormOpen}
-        menuActionType={menuActionType}
-      />
-      {isAuth &&
-        <EditMenu
-          setFormOpen={setFormOpen}
-          setMenuActionType={setMenuActionType}
-        />
-      } */}
     </>
   )
 })
